requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
cryptography==3.4.8
jsonschema==4.17.3
pytest==7.2.2
requests==2.28.2
playwright==1.32.1
pytest-playwright==0.3.2